requirements.txt
aiofiles==22.1.0
anyio==3.6.2
asn1crypto==1.5.1
attrs==22.1.0
Authlib==1.2.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
distlib==0.3.6
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.0.0
fastapi==0.85.2
filelock==3.8.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
identify==2.5.8
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
nodeenv==1.7.0
packaging==21.3
passlib==1.7.4
pg8000==1.29.3
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
scramp==1.4.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
SQLAlchemy-Utils==0.38.3
starlette==0.20.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
virtualenv==20.16.6
websockets==10.4
